requirements.txt
pandas
numpy
requests
beautifulsoup4
yfinance
supabase
boto3
openpyxl
python-dotenv
earthengine-api
python-dateutil
scikit-learn
matplotlib
seaborn
